requirements.txt
torch==2.2.0+cu118
torchvision==0.17.0+cu118
torchaudio==2.2.0+cu118
-f https://download.pytorch.org/whl/cu118

numpy==1.26.3
git+https://github.com/facebookresearch/dinov2.git

scikit-learn
opencv-python==4.8.0
tqdm
Pillow==10.2.0

matplotlib
seaborn

pandas
requests
openpyxl